Split assert class
